+++ BackEnd/src/main/java/pe/edu/upc/backend/controllers/FestividadController.java
@@ -32,18 +32,18 @@ public class FestividadController {
         }).collect(Collectors.toList());
     }
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable("id") int id) {
+    public void eliminar(@PathVariable("id") Long id) {
         fS.delete(id);
     }
     @GetMapping("/festividadporlugarturistico")
-    public List<FestividadDTO> obtenerPorLugarTuristico(@RequestParam int idLugarTuristico) {
+    public List<FestividadDTO> obtenerPorLugarTuristico(@RequestParam Long idLugarTuristico) {
         return fS.obtenerPorLugarTuristico(idLugarTuristico).stream().map(y->{
             ModelMapper m = new ModelMapper();
             return m.map(y,FestividadDTO.class);
         }).collect(Collectors.toList());
     }
     @GetMapping("/cantidadlugaresturisticosporciudad")
-    public long contarPorLugarTuristico(@RequestParam int idLugarTuristico) {
+    public long contarPorLugarTuristico(@RequestParam Long idLugarTuristico) {
         return fS.contarPorLugarTuristico(idLugarTuristico);
     }
 }
