+++ BackEnd/src/main/java/pe/edu/upc/backend/controllers/PaisController.java
@@ -32,11 +32,11 @@ public class PaisController {
         }).collect(Collectors.toList());
     }
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable("id") Integer id) {
+    public void eliminar(@PathVariable("id") Long id) {
         pS.delete(id);
     }
     @GetMapping("/{id}")
-    public PaisDTO buscarPorId(@PathVariable("id") Integer id) {
+    public PaisDTO buscarPorId(@PathVariable("id") Long id) {
         ModelMapper m = new ModelMapper();
         return m.map(pS.listById(id), PaisDTO.class);
     }
