+++ BackEnd/src/main/java/pe/edu/upc/backend/repositories/ILugarTuristicoRepository.java
@@ -8,13 +8,13 @@ import pe.edu.upc.backend.entities.LugarTuristico;
 import java.util.List;
 
 @Repository
-public interface ILugarTuristicoRepository extends JpaRepository<LugarTuristico, Integer> {
+public interface ILugarTuristicoRepository extends JpaRepository<LugarTuristico, Long> {
     // Obtener todos los lugares turisticos por id en una ciudad
     @Query("Select l from LugarTuristico l where l.idCiudad.idCiudad = :idCiudad")
-    public List<LugarTuristico> obtenerPorCiudadL(@Param("idCiudad") int idCiudad);
+    public List<LugarTuristico> obtenerPorCiudadL(@Param("idCiudad") Long idCiudad);
     // Contar cuántos lugares turisticos hay por id en una ciudad
     @Query("Select count(l) from LugarTuristico l where l.idCiudad.idCiudad = :idCiudad")
-    public long contarPorCiudadL(@Param("idCiudad") int idCiudad);
+    public long contarPorCiudadL(@Param("idCiudad") Long idCiudad);
     //Buscar numero por id
     @Query("Select lt from LugarTuristico lt where lt.nombreLugarTuristico = :nombreLugarTuristico")
     public List<LugarTuristico> listarPorNumero(@Param("nombreLugarTuristico") String nombreLugarTuristico);
