+++ BackEnd/src/main/java/pe/edu/upc/backend/repositories/IMenbresiaVIPRepository.java
@@ -3,18 +3,17 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.RequestParam;
 import pe.edu.upc.backend.entities.MenbresiaVIP;
 import java.time.LocalDate;
 import java.util.List;
 @Repository
-public interface IMenbresiaVIPRepository extends JpaRepository<MenbresiaVIP, Integer> {
+public interface IMenbresiaVIPRepository extends JpaRepository<MenbresiaVIP, Long> {
     // Obtener todas las membresías activas
     @Query("Select m from MenbresiaVIP m where m.estadoMenbresiaVIP = true")
     public List<MenbresiaVIP> obtenerMenbresActiv();
     // Obtener todas las membresías de un usuario específico
     @Query("Select m from MenbresiaVIP m where m.idUsuario.idUsuario = :idUsuario")
-    public List<MenbresiaVIP> obtenerMenbresDeUsuar(@Param("idUsuario") int idUsuario);
+    public List<MenbresiaVIP> obtenerMenbresDeUsuar(@Param("idUsuario") Long idUsuario);
     // Obtener membresías cuyo tipo sea igual a un valor específico
     @Query("Select m from MenbresiaVIP m where m.tipoMenbresiaVIP = :tipo")
     public List<MenbresiaVIP> obtenerMenbresiasTipoMayorEspecifico(@Param("tipo") String tipo);
@@ -26,11 +25,11 @@ public interface IMenbresiaVIPRepository extends JpaRepository<MenbresiaVIP, Int
     public List<MenbresiaVIP> obtenerMenbresExpiranEnFechaEspecifi(@Param("fecha") LocalDate fecha);
     // Obtener las membresías activas de un usuario específico
     @Query("Select m from MenbresiaVIP m where m.idUsuario.idUsuario = :idUsuario and m.estadoMenbresiaVIP = true")
-    public List<MenbresiaVIP> obtenerMenbresDeUsuario(@Param("idUsuario") int idUsuario);
+    public List<MenbresiaVIP> obtenerMenbresDeUsuario(@Param("idUsuario") Long idUsuario);
+    //Obtener el nombre del usuario y su tipo de membresía
     @Query(value = "SELECT mv.* \n" +
             " FROM menbresiavip mv \n" +
             " JOIN usuario u ON mv.id_usuario = u.id_usuario \n" +
             " WHERE mv.tipo_menbresiavip = :tipoMenbresiaVIP", nativeQuery = true)
     List<MenbresiaVIP> obtenerNombreUsuarioYTipoMembresia(@Param("tipoMenbresiaVIP") String tipo);
-    //Obtener el nombre del usuario y su tipo de membresía
 }
