+++ BackEnd/src/main/java/pe/edu/upc/backend/serviceimplements/LugarTuristicoImplement.java
@@ -16,13 +16,13 @@ public class LugarTuristicoImplement implements ILugarTuristicoService {
     @Override
     public void update(LugarTuristico lugarTuristico) { lR.save(lugarTuristico); }
     @Override
-    public void delete(int idLugarTuristico) { lR.deleteById(idLugarTuristico); }
+    public void delete(Long idLugarTuristico) { lR.deleteById(idLugarTuristico); }
     @Override
-    public LugarTuristico listById(int listById) { return lR.findById(listById).orElse(new LugarTuristico()); }
+    public LugarTuristico listById(Long listById) { return lR.findById(listById).orElse(new LugarTuristico()); }
     @Override
-    public List<LugarTuristico> obtenerPorCiudadL(int idCiudad) { return lR.obtenerPorCiudadL(idCiudad); }
+    public List<LugarTuristico> obtenerPorCiudadL(Long idCiudad) { return lR.obtenerPorCiudadL(idCiudad); }
     @Override
-    public long contarPorCiudadL(int idCiudad) { return lR.contarPorCiudadL(idCiudad); }
+    public long contarPorCiudadL(Long idCiudad) { return lR.contarPorCiudadL(idCiudad); }
     @Override
     public List<LugarTuristico> listarPorNumero(String nombreLugarTuristico) { return lR.listarPorNumero(nombreLugarTuristico); }
     @Override
